requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Data Processing
python-dotenv==1.0.0

# RAG and Vector Database
faiss-cpu==1.13.2
sentence-transformers==5.2.2

# EXPERIMENTAL: AI Enhancement Layer
openai==1.54.0

# Future dependencies (will be added in later tasks)
# langchain==0.0.340
# redis==5.0.1
